refactor: Update Home component to navigate to FichaUsuario page
- Import the `useNavigate` hook from `react-router-dom`
- Add `useNavigate` hook to create a navigation function
- Update the button onClick event to navigate to the `/ficha-usuario` page
